Align add registry and template prompting
