+++ packages/wp-typia/src/command-option-metadata.ts
@@ -210,7 +210,8 @@ export const ADD_OPTION_METADATA = {
     type: 'string',
   },
   template: {
-    description: 'Built-in block family for the new block.',
+    description:
+      'Optional built-in block family for the new block; interactive terminals prompt when omitted and non-interactive runs default to basic.',
     type: 'string',
   },
   to: {
