+++ packages/wp-typia/tests/cli-package.test.ts
@@ -8,6 +8,7 @@ import {
   WP_TYPIA_NODE_FALLBACK_TOP_LEVEL_COMMAND_NAMES,
   WP_TYPIA_TOP_LEVEL_COMMAND_NAMES,
 } from '../src/command-contract';
+import { formatAddKindUsagePlaceholder } from '../src/add-kind-registry';
 import {
   hasFlagBeforeTerminator,
   parseGlobalFlags,
@@ -365,8 +366,12 @@ describe('wp-typia package', () => {
     expect(addHelpOutput).toContain('--external-layer-source');
     expect(addHelpOutput).toContain('--external-layer-id');
     expect(addHelpOutput).toContain('--alternate-render-targets');
+    expect(addHelpOutput).toContain(
+      'interactive terminals prompt when omitted and non-interactive runs default to basic',
+    );
     expect(addHelpOutput).toContain('admin-view');
-    expect(addHelpOutput).toContain('ability, ai-feature');
+    expect(addHelpOutput).toContain('ability');
+    expect(addHelpOutput).toContain('ai-feature');
     expect(addHelpOutput).toContain('editor-plugin');
   });
 
@@ -1575,11 +1580,13 @@ describe('wp-typia package', () => {
     expect(result.status).toBe(1);
     expect(result.stdout).toContain('Usage:');
     expect(result.stdout).toContain('wp-typia add admin-view <name>');
-    expect(result.stdout).toContain('wp-typia add block <name>');
+    expect(result.stdout).toContain(
+      'wp-typia add block <name> [--template <basic|interactivity|persistence|compound>]',
+    );
     expect(result.stdout).toContain('wp-typia add style <name>');
     expect(result.stdout).toContain('wp-typia add transform <name>');
     expect(result.stderr).toContain(
-      '`wp-typia add` requires <kind>. Usage: wp-typia add <admin-view|block|variation|style|transform|pattern|binding-source|rest-resource|ability|ai-feature|editor-plugin|hooked-block> ...',
+      `\`wp-typia add\` requires <kind>. Usage: wp-typia add ${formatAddKindUsagePlaceholder()} ...`,
     );
   });
 
@@ -1596,7 +1603,9 @@ describe('wp-typia package', () => {
       expect(capturedStdout.join('\n')).toContain(
         'wp-typia add admin-view <name>',
       );
-      expect(capturedStdout.join('\n')).toContain('wp-typia add block <name>');
+      expect(capturedStdout.join('\n')).toContain(
+        'wp-typia add block <name> [--template <basic|interactivity|persistence|compound>]',
+      );
       expect(capturedStdout.join('\n')).toContain(
         'wp-typia add ability <name>',
       );
